- changed existing code into DynamicFormOld - added new DynamicForm which takes a form-first approach - added nextjs as a test dep
